+++ src/main/java/net/pwall/pipeline/AppendableAcceptor.java
@@ -1,8 +1,8 @@
 /*
- * @(#) BaseAcceptor.java
+ * @(#) AppendableAcceptor.java
  *
  * pipelines   Pipeline conversion library for Java
- * Copyright (c) 2020 Peter Wall
+ * Copyright (c) 2021 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -25,45 +25,24 @@
 
 package net.pwall.pipeline;
 
-/**
- * The base interface for pipeline and acceptor classes.
- *
- * @author  Peter Wall
- * @param   <R>     the result type
- */
-public interface BaseAcceptor<R> extends AutoCloseable {
+public class AppendableAcceptor<R> extends AbstractIntAcceptor<R> {
 
-    /**
-     * Return {@code true} if the pipeline is closed.
-     *
-     * @return  {@code true} if the pipeline is closed
-     */
-    boolean isClosed();
+    private final Appendable appendable;
 
-    /**
-     * Get the result of the acceptor.  The default implementation throws an exception.
-     *
-     * @return  the result
-     * @throws  UnsupportedOperationException   in all cases
-     */
-    default R getResult() {
-        throw new UnsupportedOperationException("No result possible");
+    public AppendableAcceptor(Appendable appendable) {
+        this.appendable = appendable;
     }
 
     /**
-     * Return {@code true} if all sequences in the input are complete, that is, the input is not in the middle of a
-     * sequence requiring more data.  This should be overridden by acceptors that process complex sequences of input.
+     * Accept an {@code int}, after {@code closed} check and test for end of data.  Forward the value to the
+     * {@link Appendable} as a {@code char}.
      *
-     * @return  {@code true} if the input is in the "complete" state
+     * @param   value       the input value
+     * @throws  Exception   if thrown by the {@link Appendable} or by a {@code close()} method
      */
-    default boolean isComplete() {
-        return true;
-    }
-
-    /**
-     * Flush the acceptor (required by some implementations - default is a no-operation).
-     */
-    default void flush() {
+    @Override
+    public void acceptInt(int value) throws Exception {
+        appendable.append((char)value);
     }
 
 }