+++ src/main/java/net/pwall/pipeline/codec/SwitchableDecoder.java
@@ -29,15 +29,14 @@ import java.nio.charset.Charset;
 
 import net.pwall.pipeline.AbstractIntPipeline;
 import net.pwall.pipeline.IntAcceptor;
-import net.pwall.pipeline.IntPipeline;
 
 /**
  * A switchable decoder that allows the encoding of an input stream to be determined from the data in the early part of
  * the stream.
  *
- * <p>An XML file declares its encoding on the prefix line, and an HTML can contain a META tag specifying the content
- * type and encoding of the data.  These rely on the declaration occurring early in the file, before any characters
- * outside the base ASCII set are expected to be encountered.</p>
+ * <p>An XML file declares its encoding on the prefix line, and an HTML file can contain a META tag specifying the
+ * content type and encoding of the data.  These rely on the declaration occurring early in the file, before any
+ * characters outside the base ASCII set are expected to be encountered.</p>
  *
  * <p>This decoder allows the reading of an input stream to commence in ASCII decoding, switching to the full decoding
  * once the declaration has been read.</p>
@@ -47,17 +46,15 @@ import net.pwall.pipeline.IntPipeline;
  */
 public class SwitchableDecoder<R> extends AbstractIntPipeline<R> {
 
-    private final IntAcceptor<R> downstream;
-    private IntPipeline<R> delegate;
+    protected IntAcceptor<? extends R> delegate;
 
     /**
      * Construct a {@code SwitchableDecoder} with the given downstream {@link IntAcceptor}.
      *
      * @param downstream    the {@link IntAcceptor}
      */
-    public SwitchableDecoder(IntAcceptor<R> downstream) {
+    public SwitchableDecoder(IntAcceptor<? extends R> downstream) {
         super(downstream);
-        this.downstream = downstream;
         delegate = null;
     }
 
@@ -77,21 +74,44 @@ public class SwitchableDecoder<R> extends AbstractIntPipeline<R> {
     }
 
     /**
+     * Return {@code true} if all sequences in the input to this stage of the pipeline are complete, that is, the input
+     * is not in the middle of a sequence requiring more data.
+     *
+     * @return {@code true} if the input is in the "complete" state
+     */
+    @Override
+    public boolean isStageComplete() {
+        return delegate == null || delegate.isComplete();
+    }
+
+    /**
+     * Switch to the specified delegate.
+     *
+     * @param   delegate    the new delegate
+     * @throws  Exception   if thrown by downstream
+     */
+    public void switchTo(IntAcceptor<? extends R> delegate) throws Exception {
+        this.delegate = delegate;
+    }
+
+    /**
      * Switch to the nominated character set.
      *
      * @param   charset     the {@link Charset}
+     * @throws  Exception   if thrown by downstream
      */
-    public void switchTo(Charset charset) {
-        delegate = DecoderFactory.getDecoder(charset, downstream);
+    public void switchTo(Charset charset) throws Exception {
+        switchTo(DecoderFactory.getDecoder(charset, getDownstream()));
     }
+
     /**
      * Switch to the nominated character set.
      *
      * @param   charsetName the character set name
+     * @throws  Exception   if thrown by downstream
      */
-
-    public void switchTo(String charsetName) {
-        delegate = DecoderFactory.getDecoder(charsetName, downstream);
+    public void switchTo(String charsetName) throws Exception {
+        switchTo(DecoderFactory.getDecoder(charsetName, getDownstream()));
     }
 
 }