+++ src/main/java/net/pwall/pipeline/codec/Windows1252_CodePoint.java
@@ -36,7 +36,7 @@ import net.pwall.pipeline.IntPipeline;
  */
 public class Windows1252_CodePoint<R> extends DecodingPipeline<R> {
 
-    private static final String table =
+    static final String table =
             "\u20AC\u0081\u201A\u0192\u201E\u2026\u2020\u2021\u02C6\u2030\u0160\u2039\u0152\u008D\u017D\u008F" +
             "\u0090\u2018\u2019\u201C\u201D\u2022\u2013\u2014\u02DC\u2122\u0161\u203A\u0153\u009D\u017E\u0178" +
             "\u00A0\u00A1\u00A2\u00A3\u00A4\u00A5\u00A6\u00A7\u00A8\u00A9\u00AA\u00AB\u00AC\u00AD\u00AE\u00AF" +