+++ src/test/java/net/pwall/pipeline/ListAcceptorTest.java
@@ -1,5 +1,5 @@
 /*
- * @(#) BaseAcceptor.java
+ * @(#) ListAcceptorTest.java
  *
  * pipelines   Pipeline conversion library for Java
  * Copyright (c) 2020 Peter Wall
@@ -25,45 +25,23 @@
 
 package net.pwall.pipeline;
 
-/**
- * The base interface for pipeline and acceptor classes.
- *
- * @author  Peter Wall
- * @param   <R>     the result type
- */
-public interface BaseAcceptor<R> extends AutoCloseable {
+import java.util.List;
 
-    /**
-     * Return {@code true} if the pipeline is closed.
-     *
-     * @return  {@code true} if the pipeline is closed
-     */
-    boolean isClosed();
+import org.junit.Test;
+import static org.junit.Assert.assertEquals;
 
-    /**
-     * Get the result of the acceptor.  The default implementation throws an exception.
-     *
-     * @return  the result
-     * @throws  UnsupportedOperationException   in all cases
-     */
-    default R getResult() {
-        throw new UnsupportedOperationException("No result possible");
-    }
+public class ListAcceptorTest {
 
-    /**
-     * Return {@code true} if all sequences in the input are complete, that is, the input is not in the middle of a
-     * sequence requiring more data.  This should be overridden by acceptors that process complex sequences of input.
-     *
-     * @return  {@code true} if the input is in the "complete" state
-     */
-    default boolean isComplete() {
-        return true;
-    }
-
-    /**
-     * Flush the acceptor (required by some implementations - default is a no-operation).
-     */
-    default void flush() {
+    @Test
+    public void shouldBuildListFromSequence() throws Exception {
+        ListAcceptor<String> listAcceptor = new ListAcceptor<>();
+        listAcceptor.accept("Hello");
+        listAcceptor.accept("World!");
+        listAcceptor.close();
+        assertEquals(2, listAcceptor.getSize());
+        List<String> list = listAcceptor.getResult();
+        assertEquals("Hello", list.get(0));
+        assertEquals("World!", list.get(1));
     }
 
 }